Rename cutty.{common => core}
