+++ src/cutty/core/exceptions.py
