+++ src/cutty/core/types.py
